simple guessing game c++
This code was made for my programming class and I thought I should start making a little portfolio for my self. This is made in visual studios so the code would work best there. So if there is any slopy code in there I'm still a little new to coding. and also sorry for all the blank space the teacher demands that space be there for full credit.